Serialize each arg as independent JSON, send as Form Data
This makes it a bit easier to read what is happening in dev tools, and paves the way for us also supporting file uploads in future.

It will also make interoperability with normal forms a little bit easier... literally have a HTML form and it will post to the @:serverAction with each form value mapping to an argument on the function.

Not sure how we'll map GET requests to actions at this point.