Changed TriggerActions to allow redirects after the action.
This allows you to create a new resource in an action, and then redirect to the page for that new resource.  I'm chosing to do this instead of having actions offer return values that are available client side - I feel like it keeps the architecture simple and keeps open the option for server-side action handling.

Things to note:

- TriggerAction now returns `Promsie<Noise>` rather than `Promise<TProps>`
- If an action returns a redirect, the client-side code will receive the response, log any traces to the console, and then redirect immediately - it will not attempt to re-render.
- The promise for the action will still be resolved successfully (with the value "Noise").