More code cleanup
Move a bunch of stuff from UniversalPage into static methods in SUUniversalPageClientUtils. UniversalPage is now nice and clean, which is probably what we want for a class we expect people to extend.

I've also moved SmallUniverseRoute out into it's own file.

And moved "hydrate" from `SmallUniverse.hydrate` to `UniversalPage.hydrate`.

I think there will eventually be a helper that calls this for you, so the API of calling it I'm not too worried about, and having the method sit with UniversalPage seems cleaner to me.